Сливаемый комит находится на ветке *origin/feature/new_optimization* и на данный момент является актуальной версией для
проекта *fwm-intf-module*.
Модуль [fwm-intf-module](https://gyazo.com/41012e6b5804a8358c429b1d3028dc08)
Подмодуль [Gurux.DLMS](https://gyazo.com/0ac1fd5efc0cc1152cdbec84c74ae3a4)

+++ development/include/GXDLMSActionItem.h
@@ -40,7 +40,7 @@ class CGXDLMSActionItem
 {
     friend class CGXDLMSActionSet;
     std::string m_LogicalName;
-    int m_ScriptSelector;
+    unsigned short m_ScriptSelector;
 public:
     //Constructor.
     CGXDLMSActionItem();
@@ -54,9 +54,9 @@ public:
 
     void SetLogicalName(std::string value);
 
-    int GetScriptSelector();
+	unsigned short GetScriptSelector();
 
-    void SetScriptSelector(int value);
+    void SetScriptSelector(unsigned short value);
 
     std::string ToString();
 };
